Update Materias-position.html dropzones from float to flex
Updated CSS file to change the default behacious of dropzones inside materia-edit-position.html.
Alld dropzones were changed to a flexbox model instead of a float model.
This was done to arrange dropzones in a LTR way and also to match the VLE correctly.
This change had to be done in order for the seccions to better fit on mobile devices.